refactor: Refactor get subcommand tests and improve coverage
Moved existing tests for 'get' from various files into tests/subcommands/test_get_cmd.py. Added new tests to cover missing branches in vibectl/subcommands/get_cmd.py, achieving 93% coverage. Fixed several test failures encountered during refactoring.
