+++ tests/test_cli.py
@@ -159,274 +159,6 @@ def test_cli_init_theme_error(cli_runner: CliRunner) -> None:
         assert result.exit_code == 0  # Should not fail on theme error
 
 
-def test_get_basic(
-    cli_runner: CliRunner,
-    mock_run_kubectl: Mock,
-    mock_handle_command_output: Mock,
-) -> None:
-    """Test basic get command functionality."""
-    # Note: This test mocks vibectl.command_handler.run_kubectl but in the CLI code,
-    # vibectl.command_handler.run_kubectl is actually called.
-    # We need to patch both to make this test reliable.
-    with (
-        patch("vibectl.command_handler.run_kubectl") as cmd_mock_run_kubectl,
-        patch(
-            "vibectl.command_handler.handle_command_output"
-        ) as cmd_mock_handle_output,
-        patch("vibectl.types.Success") as mock_success,
-    ):
-        # Set up mock return value to return a Success object
-        success_instance = Mock()
-        mock_success.return_value = success_instance
-        success_instance.data = "test output"
-        cmd_mock_run_kubectl.return_value = success_instance
-
-        # Make handle_command_output return a Success object
-        cmd_mock_handle_output.return_value = success_instance
-
-        # Invoke CLI
-        result = cli_runner.invoke(cli, ["get", "pods"])
-
-        # Check results - using command_handler mocks
-        assert result.exit_code == 0
-        cmd_mock_run_kubectl.assert_called_once_with(["get", "pods"], capture=True)
-        cmd_mock_handle_output.assert_called_once()
-
-
-def test_get_with_args(
-    cli_runner: CliRunner,
-    mock_run_kubectl: Mock,
-    mock_handle_command_output: Mock,
-) -> None:
-    """Test get command with additional arguments."""
-    # Need to use the correct patching strategy
-    with (
-        patch("vibectl.command_handler.run_kubectl") as cmd_mock_run_kubectl,
-        patch(
-            "vibectl.command_handler.handle_command_output"
-        ) as cmd_mock_handle_output,
-        patch("vibectl.types.Success") as mock_success,
-    ):
-        # Set up mock return value to return a Success object
-        success_instance = Mock()
-        mock_success.return_value = success_instance
-        success_instance.data = "test output"
-        cmd_mock_run_kubectl.return_value = success_instance
-
-        # Make handle_command_output return a Success object
-        cmd_mock_handle_output.return_value = success_instance
-
-        # Pass namespace as a single argument
-        result = cli_runner.invoke(cli, ["get", "pods", "-n", "default"])
-
-        # Check results - using command_handler mocks
-        assert result.exit_code == 0
-        cmd_mock_run_kubectl.assert_called_once_with(
-            ["get", "pods", "-n", "default"], capture=True
-        )
-        cmd_mock_handle_output.assert_called_once()
-
-
-def test_get_no_output(
-    cli_runner: CliRunner,
-    mock_run_kubectl: Mock,
-    mock_handle_command_output: Mock,
-) -> None:
-    """Test get command when kubectl returns no output."""
-    # Need to use the correct patching strategy
-    with (
-        patch("vibectl.command_handler.run_kubectl") as cmd_mock_run_kubectl,
-        patch(
-            "vibectl.command_handler.handle_command_output"
-        ) as cmd_mock_handle_output,
-        patch("vibectl.types.Success") as mock_success,  # Add Success mock
-    ):
-        # Set up mock return value to return a Success object with empty data
-        success_instance = Mock()
-        mock_success.return_value = success_instance
-        success_instance.data = ""  # Empty output
-        cmd_mock_run_kubectl.return_value = success_instance
-
-        # Make handle_command_output return a Success object
-        cmd_mock_handle_output.return_value = success_instance
-
-        # Invoke CLI command
-        result = cli_runner.invoke(cli, ["get", "pods"])
-
-        # Check results
-        assert result.exit_code == 0
-        cmd_mock_run_kubectl.assert_called_once_with(["get", "pods"], capture=True)
-
-
-def test_get_with_flags(
-    cli_runner: CliRunner,
-    mock_run_kubectl: Mock,
-    mock_handle_command_output: Mock,
-) -> None:
-    """Test get command with output flags."""
-    # Need to use the correct patching strategy
-    with (
-        patch("vibectl.command_handler.run_kubectl") as cmd_mock_run_kubectl,
-        patch(
-            "vibectl.command_handler.handle_command_output"
-        ) as cmd_mock_handle_output,
-        patch("vibectl.types.Success") as mock_success,  # Add Success mock
-    ):
-        # Set up mock return value to return a Success object
-        success_instance = Mock()
-        mock_success.return_value = success_instance
-        success_instance.data = "test output"
-        cmd_mock_run_kubectl.return_value = success_instance
-
-        # Make handle_command_output return a Success object
-        cmd_mock_handle_output.return_value = success_instance
-
-        result = cli_runner.invoke(
-            cli,
-            [
-                "get",
-                "pods",
-                "--show-raw-output",
-                "--no-show-vibe",
-                "--model",
-                "test-model",
-            ],
-        )
-
-        # Check results
-        assert result.exit_code == 0
-        cmd_mock_run_kubectl.assert_called_once_with(["get", "pods"], capture=True)
-
-
-def test_get_with_show_kubectl_flag(
-    cli_runner: CliRunner,
-    mock_run_kubectl: Mock,
-    mock_handle_command_output: Mock,
-) -> None:
-    """Test get command with show_kubectl flag."""
-    with (
-        patch(
-            "vibectl.subcommands.get_cmd.configure_output_flags"
-        ) as mock_configure_flags,
-        patch("vibectl.command_handler.run_kubectl") as cmd_mock_run_kubectl,
-        patch(
-            "vibectl.command_handler.handle_command_output"
-        ) as cmd_mock_handle_output,
-        patch("vibectl.types.Success") as mock_success,  # Add Success mock
-    ):
-        # Set up mock return value to return a Success object
-        success_instance = Mock()
-        mock_success.return_value = success_instance
-        success_instance.data = "test output"
-        cmd_mock_run_kubectl.return_value = success_instance
-
-        # Make handle_command_output return a Success object
-        cmd_mock_handle_output.return_value = success_instance
-
-        # Mock the configure_output_flags function
-        mock_configure_flags.return_value = Mock(
-            show_raw=False, show_vibe=True, warn_no_output=True, model_name="test-model"
-        )
-
-        result = cli_runner.invoke(
-            cli,
-            [
-                "get",
-                "pods",
-                "--show-kubectl",
-            ],
-        )
-
-        # Check results
-        assert result.exit_code == 0
-        cmd_mock_run_kubectl.assert_called_once_with(["get", "pods"], capture=True)
-
-
-def test_get_with_no_show_kubectl_flag(
-    cli_runner: CliRunner,
-    mock_run_kubectl: Mock,
-    mock_handle_command_output: Mock,
-) -> None:
-    """Test get command with no_show_kubectl flag."""
-    with (
-        patch(
-            "vibectl.subcommands.get_cmd.configure_output_flags"
-        ) as mock_configure_flags,
-        patch("vibectl.command_handler.run_kubectl") as cmd_mock_run_kubectl,
-        patch(
-            "vibectl.command_handler.handle_command_output"
-        ) as cmd_mock_handle_output,
-        patch("vibectl.types.Success") as mock_success,  # Add Success mock
-    ):
-        # Set up mock return value to return a Success object
-        success_instance = Mock()
-        mock_success.return_value = success_instance
-        success_instance.data = "test output"
-        cmd_mock_run_kubectl.return_value = success_instance
-
-        # Make handle_command_output return a Success object
-        cmd_mock_handle_output.return_value = success_instance
-
-        # Mock the configure_output_flags function
-        mock_configure_flags.return_value = Mock(
-            show_raw=False, show_vibe=True, warn_no_output=True, model_name="test-model"
-        )
-
-        result = cli_runner.invoke(
-            cli,
-            [
-                "get",
-                "pods",
-                "--no-show-kubectl",
-            ],
-        )
-
-        # Check results
-        assert result.exit_code == 0
-        cmd_mock_run_kubectl.assert_called_once_with(["get", "pods"], capture=True)
-
-
-@patch("vibectl.subcommands.get_cmd.handle_vibe_request")
-def test_get_vibe_request(mock_handle_vibe_get: Mock, cli_runner: CliRunner) -> None:
-    """Test get command with vibe request."""
-    result = cli_runner.invoke(cli, ["get", "vibe", "show", "me", "all", "pods"])
-    assert result.exit_code == 0
-    mock_handle_vibe_get.assert_called_once()
-    args, kwargs = mock_handle_vibe_get.call_args
-    assert kwargs["request"] == "show me all pods"
-    assert kwargs["command"] == "get"
-
-
-def test_get_vibe_no_request(cli_runner: CliRunner, mock_console: Mock) -> None:
-    """Test get vibe command without a request."""
-    with patch("vibectl.console.console_manager"):
-        result = cli_runner.invoke(cli, ["get", "vibe"])
-        assert result.exit_code == 1
-        assert "Error: Missing request after 'vibe'" in result.output
-
-
-def test_get_error_handling(cli_runner: CliRunner, mock_run_kubectl: Mock) -> None:
-    """Test get command error handling."""
-    # Patch handle_standard_command instead of run_kubectl to use the new Result type
-    with patch(
-        "vibectl.subcommands.get_cmd.handle_standard_command"
-    ) as mock_handle_standard:
-        # Set up mock to return an Error
-        from vibectl.types import Error
-
-        mock_error = Error(error="Test error")
-        mock_handle_standard.return_value = mock_error
-
-        # Invoke the CLI and test error handling
-        result = cli_runner.invoke(cli, ["get", "pods"])
-
-        # Verify the command was called
-        mock_handle_standard.assert_called_once()
-
-        # Check that the command failed with a non-zero exit code
-        assert result.exit_code != 0
-
-
 @patch("vibectl.cli.Config")
 def test_instructions_set_basic(mock_config_class: Mock, cli_runner: CliRunner) -> None:
     """Test basic instructions set functionality."""
